Comprehensive Documentation Update and Presentation Slides Addition (#14)
**Overview**
This pull request introduces significant updates across several
documentation and project presentation aspects. The goal is to enhance
clarity, improve user understanding, and provide a detailed walkthrough
of the project functionalities and architecture.

**Changes Included:**
1. Updated README: The README file has been thoroughly revised for
better readability and user guidance. It now includes more detailed
sections on installation, usage, and contributions, making it easier for
new users and developers to get started with the project.
2. Enhanced Docstrings: All Python scripts within the project have been
updated with comprehensive docstrings. These include detailed
descriptions of functions, parameter explanations, and information on
return types. This update aims to facilitate easier navigation and
understanding of the codebase for maintenance and future development.
3. In-Line Comments: Significant additions have been made to in-line
comments within the scripts. These comments provide insights into the
logic and decisions behind code blocks, enhancing the comprehensibility
of complex algorithms and workflows.
4. Presentation Slides: A new set of presentation slides has been added
to the repository. These slides are designed to assist in project
presentations and stakeholder meetings, offering a visual representation
of the project's scope, architecture, and the impact of recent updates.

Thank you in advance for reviewing the PR.
Samyukta

---------

Signed-off-by: Sai Samyukta Palle <141570219+pallesaisamyukta@users.noreply.github.com>